# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "resources api",
        "title": "commonpool api",
        "termsOfService": "http://swagger.io/terms",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8585",
    "basePath": "/api/v1",
    "paths": {
        "/chat/:id": {
            "post": {
                "description": "This endpoint sends a message to the given thread",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Sends a message to a topic",
                "operationId": "sendMessage",
                "parameters": [
                    {
                        "description": "Message to send",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.SendMessageRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Topic id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        },
        "/chat/messages": {
            "get": {
                "description": "This endpoint returns the messages for the given topic.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Gets topic messages",
                "operationId": "getMessages",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of messages to take",
                        "name": "take",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of messages to skip",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Topic id",
                        "name": "topic",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.GetTopicMessagesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        },
        "/chat/threads": {
            "get": {
                "description": "This endpoint returns the latest messaging threads for the currently logged in user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Returns the latest user message threads",
                "operationId": "getLatestThreads",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of threads to take",
                        "name": "take",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of threads to skip",
                        "name": "skip",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.GetLatestThreadsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        },
...
        "/resources/:id/inquire": {
            "post": {
                "description": "This endpoint sends a message to the resource owner",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Sends a message to the user about a resource",
                "operationId": "inquireAboutResource",
                "parameters": [
                    {
                        "description": "Message to send",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.InquireAboutResourceRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Resource id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "utils.Error": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
...
        "web.GetLatestThreadsResponse": {
            "type": "object",
            "properties": {
                "threads": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.Thread"
                    }
                }
            }
        },
...
        "web.GetTopicMessagesResponse": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.Message"
                    }
                }
            }
        },
        "web.InquireAboutResourceRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "web.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "sentAt": {
                    "type": "string"
                },
                "sentBy": {
                    "type": "string"
                },
                "sentByUsername": {
                    "type": "string"
                },
                "topicId": {
                    "type": "string"
                }
            }
        },
...
        "web.SendMessageRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "web.Thread": {
            "type": "object",
            "properties": {
                "hasUnreadMessages": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "lastChars": {
                    "type": "string"
                },
                "lastMessageAt": {
                    "type": "string"
                },
                "lastMessageUserId": {
                    "type": "string"
                },
                "lastMessageUsername": {
                    "type": "string"
                },
                "recipientId": {
                    "type": "string"
                }
            }
        },
...
    }
}